add basic.py
