+++ symbolic_trace/opcode_translator/executor/variables/__init__.py
@@ -1,16 +1,18 @@
 from .base import (  # noqa: F401
-    ConstantVariable,
     ConstTypes,
-    DygraphTracerVariable,
-    ModuleVariable,
-    ObjectVariable,
-    SliceVariable,
-    TensorVariable,
     VariableBase,
     VariableFactory,
     get_zero_degree_vars,
     map_variables,
     topo_sort_vars,
+)
+from .basic import (  # noqa: F401
+    ConstantVariable,
+    DygraphTracerVariable,
+    ModuleVariable,
+    ObjectVariable,
+    SliceVariable,
+    TensorVariable,
 )
 from .callable import (  # noqa: F401
     BuiltinVariable,
